re-organize registry src and test code
oops
