+++ src/registry/README.md
@@ -1,1 +1,12 @@
-[Instructions](https://github.com/alan-turing-institute/MLJ.jl/blob/master/REGISTRY.md) on registering models implementing the MLJ interface.
+The *MLJ Model Registry*, also known as the *model metadata database*,
+is the current directory called file
+[METADATA.toml](METADATA.toml). It is generated by loading all
+packages in [Project.toml](Project.toml), searching for all types
+implementing MLJ's model interface, and extracting [model
+trait](https://github.com/JuliaAI/MLJModelInterface.jl/blob/dev/src/MLJModelInterface.jl)
+values for all such types. The abbreviated list of models in
+[Models.toml](Models.toml) is generated at the same time. 
+
+
+[Instructions](https://github.com/JuliaAI/MLJModels.jl#instructions-for-updating-the-mlj-model-registry) on registering models implementing the MLJ interface.
+
