+++ src/registry/test/info_dict.jl
