+++ src/registry/test/runtests.jl
@@ -1,0 +1,5 @@
+@testset "metadata.jl" begin
+    @test include("info_dict.jl")
+end
+
+true
